Remove generator code other than enum generation
